+++ sfs-main-extension/src/com/gerantech/towers/sfs/socials/handlers/LobbyRoomServerEventsHandler.java
@@ -42,12 +42,12 @@ public class LobbyRoomServerEventsHandler extends BaseServerEventHandler
 
 			// broadcast join message
 			if( lobbyClass.getData().getMembers().size() > 1 )
-				lobbyClass.sendComment((short) MessageTypes.M10_COMMENT_JOINT, player, "", (short)-1);
+				lobbyClass.sendComment( MessageTypes.M10_COMMENT_JOINT, player, "", -1);
 		}
 		else if( arg.getType().equals(SFSEventType.USER_LEAVE_ROOM) )// mode = leave
 		{
 			// broadcast leave message
-			lobbyClass.sendComment((short) MessageTypes.M11_COMMENT_LEAVE, player, "", (short)-1);
+			lobbyClass.sendComment( MessageTypes.M11_COMMENT_LEAVE, player, "", -1);
 			LobbyUtils.getInstance().removeUser(lobbyClass.getData(), player.id);
 			LobbyUtils.getInstance().removeEmptyRoom(lobby);
